+++ examples/01-hello-world/template.tsx
@@ -1,12 +1,10 @@
 /**
- * Run with: npm run build && node dist/examples/01-hello-world/template.js
+ * Run with: npm run build && node dist/bin/fileable.js build dist/examples/01-hello-world/template.js
  */
-import { render } from "fileable";
-
 const template = (
   <dir name="dist">
     <file name="hello.txt">Hello, world!</file>
   </dir>
 );
 
-await render(template, { outDir: new URL(".", import.meta.url).pathname, cache: false });
+export default template;